+++ test/manifest/emptyPlaylistType.js
@@ -29,5 +29,6 @@
   ],
   "targetDuration": 10,
   "endList": true,
-  "discontinuitySequence": 0
+  "discontinuitySequence": 0,
+  "discontinuityStarts": []
 }