+++ test/manifest/manifestExtXEndlistEarly.js
@@ -25,5 +25,6 @@
   ],
   "targetDuration": 10,
   "endList": true,
-  "discontinuitySequence": 0
+  "discontinuitySequence": 0,
+  "discontinuityStarts": []
 }