+++ test/manifest/missingEndlist.js
@@ -12,5 +12,6 @@
     }
   ],
   "targetDuration": 10,
-  "discontinuitySequence": 0
+  "discontinuitySequence": 0,
+  "discontinuityStarts": []
 }