+++ test/manifest/negativeMediaSequence.js
@@ -22,5 +22,6 @@
   ],
   "targetDuration": 8,
   "endList": true,
-  "discontinuitySequence": 0
+  "discontinuitySequence": 0,
+  "discontinuityStarts": []
 }